requirements.txt
# .env 与模型配置
python-dotenv>=1.0.0
PyYAML>=6.0
python-frontmatter>=1.0.0
requests>=2.28.0
pytz>=2024.1

# LangChain 1.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0

# Qwen / DashScope
dashscope>=1.20.0

# 网页搜索：BochaAI MCP（中文场景首选），需 BOCHA_API_KEY
langchain-mcp-adapters>=0.1.0

# 中文分词与高频词统计
jieba>=0.42.1

# 浏览器自动化（用于 NetInsight 登录）
playwright>=1.40.0

# Postgres 连接支持
psycopg2-binary>=2.9.0

# CLI 颜色支持
colorama>=0.4.6

# CLI 框架和美化
typer>=0.9.0
rich>=13.7.0
prompt-toolkit>=3.0.43

# 测试
pytest>=8.0.0
